+++ frontend/src/components/custom/CustomRadio.tsx
@@ -5,7 +5,7 @@
 import React from 'react'
 import { useAppDispatch } from '../../store/hooks'
 import { addFilterWithRadio, deleteFilter } from '../../store/filterSlice'
-import { triggerRequest } from '../pages/requestTriggerSlice'
+import { triggerRequest } from '../../routes/requestTriggerSlice'
 
 const CustomRadio: React.FC<CustomCheckboxProps> = ({
 	id,
