+++ frontend/src/routes/namespace/index.tsx
@@ -3,33 +3,38 @@
 // SPDX-FileCopyrightText: 2019 Georg Schwarz <georg. schwarz@fau.de>
 
 import React, { useEffect, useState } from 'react'
-import { useAppDispatch, useAppSelector } from '../../../store/hooks'
+import { useAppDispatch, useAppSelector } from '../../store/hooks'
 import axios from 'axios'
-import { selectCluster, selectTenant } from '../../../store/filterSlice'
-import TenantView from './TenantView'
+import {
+	selectCluster,
+	selectNamespace,
+	selectTenant,
+} from '../../store/filterSlice'
+import NamespaceView from './NamespaceView'
 import { selectTrigger } from '../requestTriggerSlice'
-import config from '../../../config'
+import config from '../../config'
 import { Masonry } from 'react-plock'
 
-export interface ResponseTenant {
-	tenants: TenantInfo[]
+export interface ResponseNamespace {
+	namespaces: NamespaceInfo[]
 }
 
 /**
- * Card group component for the tenant type.
- * Displays the TenantView cards, title, loading window and network error.
- * @returns Rendered tenant view cards for the dashboard component
+ * Card group component for the namespace type.
+ * Displays the NamespaceView cards, title, loading window and network error.
+ * @returns Rendered namespace view cards for the dashboard component
  */
-const TenantGroup: React.FC = () => {
-	const [data, setData] = useState<TenantInfo[]>([])
+const NamespaceGroup: React.FC = () => {
+	const [data, setData] = useState<NamespaceInfo[]>([])
 	const [error, setError] = useState<string | null>(null)
 	const [loading, setLoading] = useState<boolean>(true)
 	const clusterFilter = useAppSelector(selectCluster)
 	const tenantFilter = useAppSelector(selectTenant)
-	const baseURL = config.backendUrl + '/api/tenant/all'
+	const namespaceFilter = useAppSelector(selectNamespace)
+	const baseURL = config.backendUrl + '/api/namespace/all/'
 	const trigger = useAppSelector(selectTrigger)
 
-	// Sends get request to /cluster/all for general information everytime the trigger value changes
+	// Sends get request to /namespace/all for general information everytime the trigger value changes
 	useEffect(() => {
 		// Query parameters
 		const clusterQuery = clusterFilter
@@ -38,15 +43,20 @@ const TenantGroup: React.FC = () => {
 		const tenantQuery = tenantFilter
 			.map((tenant) => `tenants=${tenant}`)
 			.join('&')
+		const namespaceQuery = namespaceFilter
+			.map((namespace) => `namespaces=${namespace}`)
+			.join('&')
 
 		// Joining all query parameters
-		const query = [clusterQuery, tenantQuery].filter((q) => q).join('&')
+		const query = [clusterQuery, tenantQuery, namespaceQuery]
+			.filter((q) => q)
+			.join('&')
 		const url = `${baseURL}?${query}`
 		// Sending GET request
 		axios
-			.get<ResponseTenant>(url)
+			.get<ResponseNamespace>(url)
 			.then((response) => {
-				setData(response.data.tenants)
+				setData(response.data.namespaces)
 				setLoading(false)
 			})
 			.catch((error) => {
@@ -57,11 +67,8 @@ const TenantGroup: React.FC = () => {
 
 	return (
 		<div>
-			<h2 className="dashboard-title">Available Tenants ({data.length})</h2>
-			<h3 className="dashboard-subtitle">
-				Namespaces: {sumElements(data, 'numberOfNamespaces')}, Topics:{' '}
-				{sumElements(data, 'numberOfTopics')}
-			</h3>
+			<h2 className="dashboard-title">Available Namespaces ({data.length})</h2>
+			<h3 className="dashboard-subtitle">Topics: {sumTopics(data)}</h3>
 			{loading ? (
 				<div className="main-card"> Loading...</div>
 			) : error ? (
@@ -75,9 +82,9 @@ const TenantGroup: React.FC = () => {
 						gap: [34, 34],
 						media: [1619, 1620],
 					}}
-					render={(tenant, index) => (
+					render={(namespace, index) => (
 						<div className="main-card" key={index}>
-							<TenantView key={index} data={tenant} />
+							<NamespaceView key={index} data={namespace} />
 						</div>
 					)}
 				/>
@@ -86,9 +93,7 @@ const TenantGroup: React.FC = () => {
 	)
 }
 
-const sumElements = (
-	tenants: TenantInfo[],
-	field: 'numberOfNamespaces' | 'numberOfTopics'
-) => tenants.map((element) => element[field]).reduce((a, b) => a + b, 0)
+const sumTopics = (namespaces: NamespaceInfo[]) =>
+	namespaces.map((element) => element.numberOfTopics).reduce((a, b) => a + b, 0)
 
-export default TenantGroup
+export default NamespaceGroup
