+++ frontend/src/routes/requestTriggerSlice.tsx
@@ -3,7 +3,7 @@
 // SPDX-FileCopyrightText: 2019 Georg Schwarz <georg. schwarz@fau.de>
 
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
-import { RootState } from '../../store'
+import { RootState } from '../store'
 
 export type RequestTriggerState = {
 	trigger: boolean
