+++ frontend/src/routes/tenant/index.tsx
@@ -3,38 +3,33 @@
 // SPDX-FileCopyrightText: 2019 Georg Schwarz <georg. schwarz@fau.de>
 
 import React, { useEffect, useState } from 'react'
-import { useAppDispatch, useAppSelector } from '../../../store/hooks'
+import { useAppDispatch, useAppSelector } from '../../store/hooks'
 import axios from 'axios'
-import {
-	selectCluster,
-	selectNamespace,
-	selectTenant,
-} from '../../../store/filterSlice'
-import NamespaceView from './NamespaceView'
+import { selectCluster, selectTenant } from '../../store/filterSlice'
+import TenantView from './TenantView'
 import { selectTrigger } from '../requestTriggerSlice'
-import config from '../../../config'
+import config from '../../config'
 import { Masonry } from 'react-plock'
 
-export interface ResponseNamespace {
-	namespaces: NamespaceInfo[]
+export interface ResponseTenant {
+	tenants: TenantInfo[]
 }
 
 /**
- * Card group component for the namespace type.
- * Displays the NamespaceView cards, title, loading window and network error.
- * @returns Rendered namespace view cards for the dashboard component
+ * Card group component for the tenant type.
+ * Displays the TenantView cards, title, loading window and network error.
+ * @returns Rendered tenant view cards for the dashboard component
  */
-const NamespaceGroup: React.FC = () => {
-	const [data, setData] = useState<NamespaceInfo[]>([])
+const TenantGroup: React.FC = () => {
+	const [data, setData] = useState<TenantInfo[]>([])
 	const [error, setError] = useState<string | null>(null)
 	const [loading, setLoading] = useState<boolean>(true)
 	const clusterFilter = useAppSelector(selectCluster)
 	const tenantFilter = useAppSelector(selectTenant)
-	const namespaceFilter = useAppSelector(selectNamespace)
-	const baseURL = config.backendUrl + '/api/namespace/all/'
+	const baseURL = config.backendUrl + '/api/tenant/all'
 	const trigger = useAppSelector(selectTrigger)
 
-	// Sends get request to /namespace/all for general information everytime the trigger value changes
+	// Sends get request to /cluster/all for general information everytime the trigger value changes
 	useEffect(() => {
 		// Query parameters
 		const clusterQuery = clusterFilter
@@ -43,20 +38,15 @@ const NamespaceGroup: React.FC = () => {
 		const tenantQuery = tenantFilter
 			.map((tenant) => `tenants=${tenant}`)
 			.join('&')
-		const namespaceQuery = namespaceFilter
-			.map((namespace) => `namespaces=${namespace}`)
-			.join('&')
 
 		// Joining all query parameters
-		const query = [clusterQuery, tenantQuery, namespaceQuery]
-			.filter((q) => q)
-			.join('&')
+		const query = [clusterQuery, tenantQuery].filter((q) => q).join('&')
 		const url = `${baseURL}?${query}`
 		// Sending GET request
 		axios
-			.get<ResponseNamespace>(url)
+			.get<ResponseTenant>(url)
 			.then((response) => {
-				setData(response.data.namespaces)
+				setData(response.data.tenants)
 				setLoading(false)
 			})
 			.catch((error) => {
@@ -67,8 +57,11 @@ const NamespaceGroup: React.FC = () => {
 
 	return (
 		<div>
-			<h2 className="dashboard-title">Available Namespaces ({data.length})</h2>
-			<h3 className="dashboard-subtitle">Topics: {sumTopics(data)}</h3>
+			<h2 className="dashboard-title">Available Tenants ({data.length})</h2>
+			<h3 className="dashboard-subtitle">
+				Namespaces: {sumElements(data, 'numberOfNamespaces')}, Topics:{' '}
+				{sumElements(data, 'numberOfTopics')}
+			</h3>
 			{loading ? (
 				<div className="main-card"> Loading...</div>
 			) : error ? (
@@ -82,9 +75,9 @@ const NamespaceGroup: React.FC = () => {
 						gap: [34, 34],
 						media: [1619, 1620],
 					}}
-					render={(namespace, index) => (
+					render={(tenant, index) => (
 						<div className="main-card" key={index}>
-							<NamespaceView key={index} data={namespace} />
+							<TenantView key={index} data={tenant} />
 						</div>
 					)}
 				/>
@@ -93,7 +86,9 @@ const NamespaceGroup: React.FC = () => {
 	)
 }
 
-const sumTopics = (namespaces: NamespaceInfo[]) =>
-	namespaces.map((element) => element.numberOfTopics).reduce((a, b) => a + b, 0)
+const sumElements = (
+	tenants: TenantInfo[],
+	field: 'numberOfNamespaces' | 'numberOfTopics'
+) => tenants.map((element) => element[field]).reduce((a, b) => a + b, 0)
 
-export default NamespaceGroup
+export default TenantGroup
