+++ frontend/src/routes/topic/index.tsx
@@ -3,7 +3,7 @@
 // SPDX-FileCopyrightText: 2019 Georg Schwarz <georg. schwarz@fau.de>
 
 import React, { useEffect, useState } from 'react'
-import { useAppDispatch, useAppSelector } from '../../../store/hooks'
+import { useAppDispatch, useAppSelector } from '../../store/hooks'
 import axios from 'axios'
 import {
 	selectCluster,
@@ -12,10 +12,10 @@ import {
 	selectSubscription,
 	selectTenant,
 	selectTopic,
-} from '../../../store/filterSlice'
+} from '../../store/filterSlice'
 import TopicView from './TopicView'
 import { selectTrigger } from '../requestTriggerSlice'
-import config from '../../../config'
+import config from '../../config'
 import { Masonry } from 'react-plock'
 import { Pagination } from '@mui/material'
 import { Box } from '@mui/system'
